ViRendered
More cleanup to ViRendered, moved it to 'm_' on the naming.

+++ api/tachyon/scene/Scene.hpp
@@ -20,7 +20,7 @@ namespace yq {
             std::vector<RenderedPtr>    things;
             std::optional<RGB3F>        background;
             // eventually things like lights, background/layers/materials/etc will go here
-            float                       utime   = 0.; // for the push constant
+            double                      utime   = 0.; // for the push constant
         };
     }
 }
